// apps/api/src/generated/kysely/valuations/ValuationsSchema.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { default as FundingChangelogTable } from './FundingChangelog';
import type { default as OutboundDeliveryTable } from './OutboundDelivery';
import type { default as PriceTable } from './Price';
import type { default as WorkerHeartbeatTable } from './WorkerHeartbeat';
import type { default as InvestmentTable } from './Investment';
import type { default as AuditLogTable } from './AuditLog';
import type { default as InvestmentAttributionTable } from './InvestmentAttribution';
import type { default as ValuationsChangeOutboxTable } from './ValuationsChangeOutbox';
import type { default as AssetTable } from './Asset';
import type { default as AssetTransferTable } from './AssetTransfer';
import type { default as LegalEntityTable } from './LegalEntity';
import type { default as FundingChangelogFundTable } from './FundingChangelogFund';
import type { default as CurrencyAssetTable } from './CurrencyAsset';
import type { default as ExchangeRateTable } from './ExchangeRate';
import type { default as EventTable } from './Event';
import type { default as WebhookSubscriptionTable } from './WebhookSubscription';
import type { default as TeamSettingsTable } from './TeamSettings';
import type { default as NoteTable } from './Note';
import type { default as TransactionTable } from './Transaction';

export default interface ValuationsSchema {
  funding_changelog: FundingChangelogTable;

  outbound_delivery: OutboundDeliveryTable;

  price: PriceTable;

  worker_heartbeat: WorkerHeartbeatTable;

  investment: InvestmentTable;

  audit_log: AuditLogTable;

  investment_attribution: InvestmentAttributionTable;

  valuations_change_outbox: ValuationsChangeOutboxTable;

  asset: AssetTable;

  asset_transfer: AssetTransferTable;

  legal_entity: LegalEntityTable;

  funding_changelog_fund: FundingChangelogFundTable;

  currency_asset: CurrencyAssetTable;

  exchange_rate: ExchangeRateTable;

  event: EventTable;

  webhook_subscription: WebhookSubscriptionTable;

  team_settings: TeamSettingsTable;

  note: NoteTable;

  transaction: TransactionTable;
}